create NoLoginsView file
