+++ Client/Frontend/Login Management/NoLoginsView.swift
@@ -1,0 +1,46 @@
+//
+//  NoLoginsView.swift
+//  Client
+//
+//  Created by Vanna Phong on 6/11/20.
+//  Copyright © 2020 Mozilla. All rights reserved.
+//
+
+import Foundation
+
+/// Empty state view when there is no logins to display.
+class NoLoginsView: UIView {
+
+    // We use the search bar height to maintain visual balance with the whitespace on this screen. The
+    // title label is centered visually using the empty view + search bar height as the size to center with.
+    var searchBarHeight: CGFloat = 0 {
+        didSet {
+            setNeedsUpdateConstraints()
+        }
+    }
+
+    lazy var titleLabel: UILabel = {
+        let label = UILabel()
+        label.font = LoginListViewModel.LoginListUX.NoResultsFont
+        label.textColor = LoginListViewModel.LoginListUX.NoResultsTextColor
+        label.text = NSLocalizedString("No logins found", tableName: "LoginManager", comment: "Label displayed when no logins are found after searching.")
+        return label
+    }()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        addSubview(titleLabel)
+    }
+
+    internal override func updateConstraints() {
+        super.updateConstraints()
+        titleLabel.snp.remakeConstraints { make in
+            make.centerX.equalTo(self)
+            make.centerY.equalTo(self).offset(-(searchBarHeight / 2))
+        }
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
